[Android] CachedFlag was moved to //base
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/db3e1d9b131df5a2e2a34114a76a0cbafc3c14d5

[Android] Move CachedFlag and CachedFieldTrialParameter to //base

Move all generic flags classes, including CachedFlag, PostNativeFlag,
CachedFieldTrialParameter and subclasses to //base/cached_flags, along
with tests.

Bug: 1442347
