+++ android/java/apk_for_test.flags
@@ -683,12 +683,12 @@
     public <init>(...);
 }
 
--keep class org.chromium.chrome.browser.flags.CachedFlag {
+-keep class org.chromium.base.cached_flags.CachedFlag {
     *** mDefaultValue;
     <init>(...);
 }
 
--keep class org.chromium.chrome.browser.flags.BraveCachedFlag {
+-keep class org.chromium.base.cached_flags.BraveCachedFlag {
     <init>(...);
 }
