+++ android/javatests/org/chromium/chrome/browser/BytecodeTest.java
@@ -28,6 +28,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import org.chromium.base.Callback;
+import org.chromium.base.FeatureMap;
 import org.chromium.base.jank_tracker.JankTracker;
 import org.chromium.base.shared_preferences.PreferenceKeyRegistry;
 import org.chromium.base.shared_preferences.SharedPreferencesManager;
@@ -333,8 +334,8 @@ public class BytecodeTest {
         Assert.assertTrue(classExists("org/chromium/chrome/browser/tasks/ReturnToChromeUtil"));
         Assert.assertTrue(classExists("org/chromium/chrome/browser/IntentHandler"));
         Assert.assertTrue(classExists("org/chromium/chrome/browser/BraveIntentHandler"));
-        Assert.assertTrue(classExists("org/chromium/chrome/browser/flags/CachedFlag"));
-        Assert.assertTrue(classExists("org/chromium/chrome/browser/flags/BraveCachedFlag"));
+        Assert.assertTrue(classExists("org/chromium/base/cached_flags/CachedFlag"));
+        Assert.assertTrue(classExists("org/chromium/base/cached_flags/BraveCachedFlag"));
         Assert.assertTrue(classExists("org/chromium/chrome/browser/logo/LogoMediator"));
         Assert.assertTrue(classExists("org/chromium/chrome/browser/logo/BraveLogoMediator"));
         Assert.assertTrue(
@@ -1022,8 +1023,9 @@ public class BytecodeTest {
                         "org/chromium/chrome/browser/BraveAppHooks"));
         Assert.assertTrue(
                 constructorsMatch(
-                        "org/chromium/chrome/browser/flags/CachedFlag",
-                        "org/chromium/chrome/browser/flags/BraveCachedFlag",
+                        "org/chromium/base/cached_flags/CachedFlag",
+                        "org/chromium/base/cached_flags/BraveCachedFlag",
+                        FeatureMap.class,
                         String.class,
                         boolean.class));
 
@@ -1452,10 +1454,12 @@ public class BytecodeTest {
                 "org/chromium/chrome/browser/bookmarks/BookmarkManagerMediator", "mContext"));
         Assert.assertTrue(fieldExists(
                 "org/chromium/chrome/browser/bookmarks/BookmarkToolbarCoordinator", "mToolbar"));
-        Assert.assertTrue(fieldExists("org/chromium/chrome/browser/bookmarks/BookmarkPage",
-                "mBookmarkManagerCoordinator"));
         Assert.assertTrue(
-                fieldExists("org/chromium/chrome/browser/flags/CachedFlag", "mDefaultValue"));
+                fieldExists(
+                        "org/chromium/chrome/browser/bookmarks/BookmarkPage",
+                        "mBookmarkManagerCoordinator"));
+        Assert.assertTrue(
+                fieldExists("org/chromium/base/cached_flags/CachedFlag", "mDefaultValue"));
         Assert.assertFalse(
                 fieldExists("org/chromium/chrome/browser/tasks/tab_groups/TabGroupModelFilter",
                         "mIsResetting"));
