+++ base/brave_base_cached_flags_java_sources.gni
@@ -1,0 +1,9 @@
+# Copyright (c) 2024 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+# Since CachedFlag that we extend and ChromeFeatureList that we patch are parts of the same
+# component, we have to embed our class to this component. Otherwise we will have cycled
+# dependency.
+brave_base_cached_flags_java_sources = [ "//brave/base/android/java/src/org/chromium/base/cached_flags/BraveCachedFlag.java" ]
