--- patches/chrome-browser-flags-BUILD.gn.patch
@@ -1,12 +1,0 @@
-diff --git a/chrome/browser/flags/BUILD.gn b/chrome/browser/flags/BUILD.gn
-index aec877a34789bb5361bf44421de8c8c9c66f4be9..544a71dfd35c258ad1e9b457a40006139957e373 100644
---- a/chrome/browser/flags/BUILD.gn
-+++ b/chrome/browser/flags/BUILD.gn
-@@ -33,6 +33,7 @@ android_library("java") {
-     "//third_party/androidx:androidx_annotation_annotation_java",
-     "//third_party/jni_zero:jni_zero_java",
-   ]
-+  sources += brave_flags_java_sources
-   srcjar_deps = [
-     ":chrome_android_java_switches_srcjar",
-     ":chrome_browser_flags_enums_srcjar",
